+++ lib/src/Pos/Node/API.hs
@@ -666,14 +666,12 @@ instance BuildableSafeGen NodeSettings where
 
 
 type SettingsAPI =
-    Tag "Settings" 'NoTagDescription
-        :> "node-settings"
+        "node-settings"
         :> Summary "Retrieves the static settings for this node."
         :> Get '[ValidJSON] (APIResponse NodeSettings)
 
 type InfoAPI =
-        Tag "Info" 'NoTagDescription
-            :> "node-info"
+            "node-info"
             :> Summary "Retrieves the dynamic information for this node."
             :> CustomQueryFlag "force_ntp_check" ForceNtpCheck
             :> Get '[ValidJSON] (APIResponse NodeInfo)
@@ -683,8 +681,8 @@ instance HasCustomQueryFlagDescription "force_ntp_check" where
 
 -- The API definition is down here for now due to TH staging restrictions. Will
 -- relocate other stuff into it's own module when the extraction is complete.
-type API =
-        SettingsAPI
+type API = (Tag "Node Monitoring API" 'NoTagDescription) :>
+        (SettingsAPI
     :<|>
         InfoAPI
     :<|>
@@ -695,6 +693,7 @@ type API =
         Summary "Restart the underlying node software."
         :> "restart-node"
         :> Post '[ValidJSON] NoContent
+        )